#5 Rename module (exceptions → utils)
