+++ xarray_units/__init__.py
@@ -1,8 +1,8 @@
-__all__ = ["exceptions", "methods", "operators"]
+__all__ = ["methods", "operators", "utils"]
 __version__ = "0.1.0"
 
 
 # submodules
-from . import exceptions
 from . import methods
 from . import operators
+from . import utils
